[REFACTOR] Move resolve_listen_ctl_addr into ListenCtlAddr
Really, this should be a `FromStr` implementation. However, we can't
have that right now because `env_config_socketaddr!` macro creates a
more general one.

Future work on that macro would allow us to more accurately model
things. But for now, it's better to locate this logic within the type
itself, rather than have it be a free-floating function.
