+++ components/common/src/error.rs
@@ -51,6 +51,7 @@ pub enum Error {
     PermissionFailed(String),
     /// When an error occurs serializing rendering context
     RenderContextSerialization(serde_json::Error),
+    RemoteSupResolutionError(String, io::Error),
     RootRequired,
     StatusFileCorrupt(PathBuf),
     StrFromUtf8Error(str::Utf8Error),
@@ -137,6 +138,10 @@ impl fmt::Display for Error {
             Error::PermissionFailed(ref e) => e.to_string(),
             Error::RenderContextSerialization(ref e) => {
                 format!("Unable to serialize rendering context, {}", e)
+            }
+            Error::RemoteSupResolutionError(ref sup_addr, ref err) => {
+                format!("Failed to resolve remote supervisor '{}': {}",
+                        sup_addr, err,)
             }
             Error::RootRequired => {
                 "Root or administrator permissions required to complete operation".to_string()
